Separates code from main into other files
