+++ src/interval.cpp
@@ -1,0 +1,4 @@
+#include "interval.hpp"
+
+const Interval Interval::empty = Interval(+infinity, - infinity);
+const Interval Interval::universe = Interval(-infinity, +infinity);
