+++ src/interval.hpp
@@ -1,4 +1,5 @@
 #pragma once
+
 #include "util.hpp"
 
 class Interval {
@@ -25,5 +26,3 @@ public:
     static const Interval empty, universe;
 };
 
-const Interval Interval::empty = Interval(+infinity, - infinity);
-const Interval Interval::universe = Interval(-infinity, +infinity);
